Match and link all remaining stream classes
Former-commit-id: 55617d2753c4fce8712586c8e3418ff6865e49a3

+++ include/Kyoto/Streams/CMemoryInStream.hpp
@@ -13,7 +13,7 @@ public:
   CMemoryInStream(const void* ptr, unsigned long len);
   CMemoryInStream(const void* ptr, unsigned long len, EOwnerShip ownership);
   virtual ~CMemoryInStream() override {}
-  virtual uint Read(void* dest, uint len) override;
+  virtual size_t Read(void* dest, size_t len) override;
 };
 
 #endif // _CMEMORYINSTREAM
